requirements.txt
alibabacloud_cdn20180510>=2.0.0
alibabacloud_dcdn20180115>=2.0.0
alibabacloud_cas20200407>=1.0.0
alibabacloud_tea_openapi>=0.3.0
alibabacloud_tea_util>=0.3.0
python-dotenv>=0.20.0
qiniu>=7.0.0
requests>=2.28.0
